+++ frontend/src/pages/SignUpPage/components/EmailInput/index.stories.tsx
