+++ frontend/src/pages/SignUpPage/components/IDInput/index.tsx
@@ -1,39 +1,32 @@
-import React, { Dispatch, SetStateAction, useContext, useEffect, useLayoutEffect } from 'react';
+import React, { Dispatch, SetStateAction, useContext } from 'react';
 
 import InputBox from '@/components/@shared/InputBox';
 import { useInput } from '@/components/@shared/InputBox/useInput';
 
 import SnackbarContext from '@/context/Snackbar';
 
-import useNicknameCheck from '@/hooks/queries/member/useNicknameCheck';
+import useIdCheck from '@/hooks/queries/member/useIDCheck';
 
-import { isValidNickname } from '@/utils/regExp';
+import { isValidID } from '@/utils/regExp';
 
-import * as Styled from '../index.styles';
+import * as Styled from '../../index.styles';
 
-interface NicknameInputProps extends ReturnType<typeof useInput> {
+interface IDInputProps extends ReturnType<typeof useInput> {
   isAnimationActive: boolean;
   setIsAnimationActive: Dispatch<SetStateAction<boolean>>;
 }
 
-const NicknameInput = ({
-  value,
-  setValue,
-  error,
-  setError,
-  isAnimationActive,
-  setIsAnimationActive,
-}: NicknameInputProps) => {
+const IDInput = ({ value, setValue, error, setError, isAnimationActive, setIsAnimationActive }: IDInputProps) => {
   const { showSnackbar } = useContext(SnackbarContext);
-  const { refetch } = useNicknameCheck({
+  const { refetch } = useIdCheck({
     storeCode: [value],
     options: {
       onSuccess: data => {
         if (data) {
-          showSnackbar('사용할 수 있는 닉네임입니다.');
+          showSnackbar('사용할 수 있는 아이디입니다.');
         }
         if (!data) {
-          setError('중복된 닉네임입니다.');
+          setError('중복된 아이디입니다.');
           setIsAnimationActive(true);
         }
       },
@@ -45,17 +38,14 @@ const NicknameInput = ({
     },
   });
 
-  useLayoutEffect(() => {
-    if (!value) {
-      return;
+  const handleChangeIDInput = (e: React.ChangeEvent<HTMLInputElement>) => {
+    if (!isValidID(value)) {
+      setError('아이디는 4자에서 16자 사이입니다.');
     }
-    if (!isValidNickname(value)) {
-      setError('닉네임는 4자에서 16자 사이입니다.');
-    }
-    if (isValidNickname(value)) {
+    if (isValidID(value)) {
       setError('');
     }
-  }, [value]);
+  };
 
   const handleIDCheckForm = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
@@ -67,9 +57,10 @@ const NicknameInput = ({
       <Styled.InputForm onSubmit={handleIDCheckForm}>
         <InputBox.Input
           handleInvalid={() => {
-            setError('닉네임를 입력해주세요');
+            setError('아이디를 입력해주세요');
           }}
-          placeholder="닉네임"
+          placeholder="아이디"
+          onChange={handleChangeIDInput}
           isAnimationActive={isAnimationActive}
           setIsAnimationActive={setIsAnimationActive}
           required
@@ -81,4 +72,4 @@ const NicknameInput = ({
   );
 };
 
-export default NicknameInput;
+export default IDInput;
