+++ frontend/src/pages/SignUpPage/components/NicknameInput/index.stories.tsx
