+++ frontend/src/pages/SignUpPage/components/PasswordConfirmationInput/index.stories.tsx
