+++ frontend/src/pages/SignUpPage/components/PasswordConfirmationInput/index.tsx
@@ -5,26 +5,27 @@ import { useInput } from '@/components/@shared/InputBox/useInput';
 
 import { isValidPassword } from '@/utils/regExp';
 
-import * as Styled from '../index.styles';
+import * as Styled from '../../index.styles';
 
-interface PasswordInputProps extends ReturnType<typeof useInput> {
+interface PasswordConfirmationInputProps extends ReturnType<typeof useInput> {
   isAnimationActive: boolean;
   setIsAnimationActive: Dispatch<SetStateAction<boolean>>;
 }
 
-const PasswordInput = ({
+const PasswordConfirmationInput = ({
   value,
   setValue,
   error,
   setError,
   isAnimationActive,
   setIsAnimationActive,
-}: PasswordInputProps) => {
+}: PasswordConfirmationInputProps) => {
   const handleBlur = () => {
     if (!isValidPassword(value)) {
       setError('비밀번호는 영어,숫자,특수문자가 포함되어 8자에서 20자 이하입니다.');
     }
   };
+
   return (
     <InputBox value={value} setValue={setValue} error={error} setError={setError}>
       <Styled.InputForm as="div">
@@ -34,7 +35,7 @@ const PasswordInput = ({
             handleInvalid={() => {
               setError('비밀번호를 입력해주세요');
             }}
-            placeholder="비밀번호"
+            placeholder="비밀번호 확인"
             isAnimationActive={isAnimationActive}
             setIsAnimationActive={setIsAnimationActive}
             onBlur={handleBlur}
@@ -47,4 +48,4 @@ const PasswordInput = ({
   );
 };
 
-export default PasswordInput;
+export default PasswordConfirmationInput;
