+++ frontend/src/pages/SignUpPage/components/PasswordInput/index.stories.tsx
