+++ frontend/src/pages/SignUpPage/components/PasswordInput/index.tsx
@@ -5,27 +5,26 @@ import { useInput } from '@/components/@shared/InputBox/useInput';
 
 import { isValidPassword } from '@/utils/regExp';
 
-import * as Styled from '../index.styles';
+import * as Styled from '../../index.styles';
 
-interface PasswordConfirmationInputProps extends ReturnType<typeof useInput> {
+interface PasswordInputProps extends ReturnType<typeof useInput> {
   isAnimationActive: boolean;
   setIsAnimationActive: Dispatch<SetStateAction<boolean>>;
 }
 
-const PasswordConfirmationInput = ({
+const PasswordInput = ({
   value,
   setValue,
   error,
   setError,
   isAnimationActive,
   setIsAnimationActive,
-}: PasswordConfirmationInputProps) => {
+}: PasswordInputProps) => {
   const handleBlur = () => {
     if (!isValidPassword(value)) {
       setError('비밀번호는 영어,숫자,특수문자가 포함되어 8자에서 20자 이하입니다.');
     }
   };
-
   return (
     <InputBox value={value} setValue={setValue} error={error} setError={setError}>
       <Styled.InputForm as="div">
@@ -35,7 +34,7 @@ const PasswordConfirmationInput = ({
             handleInvalid={() => {
               setError('비밀번호를 입력해주세요');
             }}
-            placeholder="비밀번호 확인"
+            placeholder="비밀번호"
             isAnimationActive={isAnimationActive}
             setIsAnimationActive={setIsAnimationActive}
             onBlur={handleBlur}
@@ -48,4 +47,4 @@ const PasswordConfirmationInput = ({
   );
 };
 
-export default PasswordConfirmationInput;
+export default PasswordInput;
